Final commit
I'm sorry about this, but I've hit a wall converting the NFA to a DFA. I've been grinding away at it for days and I just can't get it to work properly, I'm sorry. It can form the NFA and start to form the states of the DFA, but I can't get it to keep going until it's completed and figure out the transitions.